add utilities file and doc content
