+++ src/pidgy/weave.py
@@ -1,6 +1,5 @@
 from typing import Any
-
-from . import get_ipython
+from .utils import get_ipython, is_widget
 from .models import Weave, dataclass, field
 
 __all__ = ("Weave",)
@@ -33,6 +32,7 @@ class Weave(Weave):
     outputs: dict = field(dict)
     shell: Any = field(get_ipython, "the current interactive shell")
     environment: Any = field(get_environment, "a jinja templating environment")
+    iframe_height: int = field(600, "the height of an iframe when urls are found")
 
     def use_asynch(self, input=True):
         self.asynch = input
@@ -53,6 +53,8 @@ class Weave(Weave):
 
         def normalize(self, type, object, metadata) -> str:
             """normalize and object with (mime)type and return a string."""
+            from .utils import get_minified, get_decoded
+
             if type == "text/html" or "svg" in type:
                 object = get_minified(object)
 
@@ -66,8 +68,9 @@ class Weave(Weave):
             return object
 
         def __call__(self, object):
-            """try to convert the object into a markdown or html
-            representation using the current ipython shell."""
+            """convert an object into a markdown/html representation"""
+            from .utils import get_active_types
+
             datum = get_ipython().display_formatter.format(object)
             data, metadata = datum if isinstance(datum, tuple) else (datum, {})
             try:
@@ -79,21 +82,24 @@ class Weave(Weave):
             return self.normalize(key, data[key], metadata)
 
     def post_run_cell(self, result):
+        from .utils import Output
         from IPython.display import display
 
         if result.error_in_exec or result.error_before_exec:
-            pass
+            pass  # don't do anything when there are errors
         elif not self.no_show.match(result.info.raw_cell):
             metadata = self.shell.kernel.get_parent().get("metadata", {})
             id = metadata.get("cellId")
             output = self.outputs.setdefault(
-                id, self.Output(parent=self, cell_id=id, display_cls=self.display_cls)
+                id, Output(parent=self, cell_id=id, display_cls=self.display_cls)
             )
             output.set_input(result.info.raw_cell)
             display(output)
             self.vars.update(output.vars)
 
     def pre_execute(self):
+        from .utils import is_widget
+
         self.vars.clear()
         self.vars.update(*(x.vars for x in self.outputs.values()))
 
@@ -152,84 +158,6 @@ class Weave(Weave):
         self.shell.events.unregister("post_run_cell", self.post_run_cell)
         self.shell.events.unregister("post_execute", self.post_execute)
 
-    @dataclass
-    class Output:
-        parent: Any
-        display_cls: Any = None
-        input: str = ""
-        display_handle: Any = None
-        cell_id: str = None
-        vars: set = field(set)
-
-        def set_input(self, input):
-            from jinja2.meta import find_undeclared_variables
-
-            self.input = input
-            self.vars.clear()
-            vars = find_undeclared_variables(self.parent.environment.parse(input))
-            self.vars.update(vars)
-            primary_loader = self.parent.environment.loader.loaders[0]
-            primary_loader.mapping.update({self.cell_id: input})
-
-        def _ipython_display_(self):
-            from IPython.core.display import DisplayHandle
-
-            if self.input.startswith(("http://", "https://")):
-                lines = self.input.splitlines()
-                if all(
-                    line.startswith(("http://", "https://"))
-                    for line in lines
-                    if line.strip()
-                ):
-                    from IPython.display import IFrame, Image, display
-                    from mimetypes import guess_type
-
-                    displays = []
-                    for line in lines:
-                        if line.strip():
-                            type, _ = guess_type(line)
-                            if type and type.startswith(("image/",)):
-                                displays.append(Image(url=line))
-                            else:
-                                displays.append(IFrame(line, height=600, width="100%"))
-                    return display(*displays)
-
-            if self.display_handle is None:
-                self.display_handle = DisplayHandle()
-            if self.parent.reactive:
-                self.display_handle.display(self.display_cls(""))
-                if self.parent.asynch:
-                    from asyncio import ensure_future
-
-                    ensure_future(self.aupdate())
-                else:
-                    self.update()
-            else:
-                self.display()
-
-        @property
-        def template(self):
-            if self.cell_id:
-                return self.parent.environment.get_template(self.cell_id)
-            return self.parent.environment.from_string(self.input)
-
-        def update(self):
-            self.display_handle.update(
-                self.display_cls(self.template.render(**self.parent.get_ns()))
-            )
-
-        def display(self):
-            self.display_handle.display(
-                self.display_cls(self.template.render(**self.parent.get_ns()))
-            )
-
-        async def aupdate(self):
-            self.display_handle.update(
-                self.display_cls(
-                    await self.template.render_async(**self.parent.get_ns())
-                )
-            )
-
 
 def load_ipython_extension(shell):
     from IPython.display import Markdown
@@ -248,38 +176,3 @@ def load_ipython_extension(shell):
 
 def unload_ipython_extension(shell):
     shell.has_trait("weave") and shell.weave.unload()
-
-
-def get_active_types(shell=None):
-    """get the active types in the current IPython shell.
-    we ignore latex, but i forget why."""
-    shell = shell or get_ipython()
-    if shell:
-        object = list(shell.display_formatter.active_types)
-        object.insert(object.index("text/html"), object.pop(object.index("text/latex")))
-        return reversed(object)
-    return []
-
-
-def get_minified(x):
-    return __import__("htmlmin").minify(x, False, True, True, True, True, True, True)
-
-
-def get_decoded(object):
-    if isinstance(object, bytes):
-        object = __import__("base64").b64encode(object).decode("utf-8")
-    return object
-
-
-def is_widget(object):
-    """is an object a widget"""
-    from sys import modules
-
-    if "ipywidgets" in modules:
-        return isinstance(object, __import__("ipywidgets").Widget)
-    return False
-
-
-def was_displayed(object):
-    # the best we can know is if the widget was ever displayed
-    return object._trait_values.get("_display_callbacks") is not None
